server/requirements.txt
alembic==1.14.1
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.11.12
click==8.1.8
cloudinary==1.44.1
Faker==37.0.2
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.11
python-dotenv==1.0.1
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
Werkzeug==3.1.3
